deduplicate functional options between the unix and windows versions of the client
